cmd/govim/internal/types: refactor into multiple files
Makes a bit more sense to have separate files for the major types in
this package that have multiples methods, supporting functions etc
